// src/generated/prisma/enums.ts
export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  CHURCH_ADMIN: 'CHURCH_ADMIN',
  PASTOR: 'PASTOR',
  DEPARTMENT_LEADER: 'DEPARTMENT_LEADER',
  MEMBER: 'MEMBER'
} as const

export type Role = (typeof Role)[keyof typeof Role]